Changing seat signals from sensors to actuators
As of today seat settings can only be set by switches, making it
difficult to e.g. restore an old setting. By changing sensors to
actuators it will be easy for e.g. a VSS client to e.g. restore a seat
to a previously used setting.

Also added a signal for seat height, previously only switches existed
for that setting, not any sensor/actuator.

Fixes #315