Added Signal based on discussion
This adds basic support for Driver Monitoring Systems (DMS) to the
ADAS spec. These have been derived from the other ADAS features and
presently include:

- DMS.IsEnabled to determine if the system has been enabled or not
- DMS.IsWarning to determine if the system is in an alert state
- DMS.IsError to determine if the system is in an error state

The DMS.IsEnabled signal can be used to determine whether driver
monitoring signals under Driver are expected to be valid, while
the DMS.IsWarning signal can be used as a bassis for interrogating
the cause of a specific alert condition from the Driver state.

Signed-off-by: Paul Mundt <paul.mundt@harman.com>
Signed-off-by: Umang Sharma <umang.sharma@gmobis.com>