Clarify VSS expectations on read/write
I ended up in a discussion of what "doing a read" in VSS means.
In short, as of today that we say that a read returns actual value.
This PR suggests to change the wording somewhat, so that a read unless otherwise specified
shall return the actual value.

I.e. it is not forbidden by VSS (project) to have a read-function that returns the desired value instead,
but then that must be clearly specified in the corresponding implementation specification.

Signed-off-by: Erik Jaegervall <erik.jaegervall@se.bosch.com>