[#822] Change granularity of TraveledDistance data entry (#824)
The TraveledDistance and its accompanying data entries
(TraveledDistanceSinceStart, TripMeterReading) have been defined
using kilometer granularity.

However, use cases in the commercial vehicle domain often require
a granularity down to meters. The existing definition using the float
data type theoretically also supports specifying distances with meter
granularity, but does so only with the usual shortcomings of encoding
float values in computer systems (rounding issues etc)

Changing the unit to meters and the data type to uint32 would allow
the usage of the data entries in the use cases mentioned above, i.e.
without the need to create an overlay for adapting them, and would
also remove the possibility to have negative values, which do not
seem to be consistent with semantics of these data entries.

Signed-off-by: Kai Hudalla <kai.hudalla@etas.com>